Move mocks in single package
