+++ code/go/0chain.net/chaincore/block/block_summary_entity_test.go
@@ -12,7 +12,7 @@ import (
 	"0chain.net/core/encryption"
 	"0chain.net/core/memorystore"
 	"0chain.net/core/util"
-	mocks "0chain.net/mocks/core/datastore"
+	"0chain.net/mocks"
 )
 
 func init() {
