+++ code/go/0chain.net/mocks/batch.go
