+++ code/go/0chain.net/mocks/iterator.go
