+++ code/go/0chain.net/mocks/query.go
