+++ code/go/0chain.net/mocks/store.go
